+++ sdk/appcenter-distribute/src/main/java/com/microsoft/appcenter/distribute/ReleaseDownloadListener.java
@@ -13,6 +13,8 @@ import androidx.annotation.Nullable;
 import androidx.annotation.UiThread;
 import androidx.annotation.WorkerThread;
 
+import android.content.Intent;
+import android.net.Uri;
 import android.widget.Toast;
 
 import com.microsoft.appcenter.distribute.download.ReleaseDownloader;
@@ -96,8 +98,11 @@ class ReleaseDownloadListener implements ReleaseDownloader.Listener {
             @Override
             public void run() {
 
+                /* Use intent to just resume app instead on install intent. */
+                Intent intent = DistributeUtils.getResumeAppIntent(mContext);
+
                 /* Check if app should install now. */
-                if (!Distribute.getInstance().notifyDownload(mReleaseDetails)) {
+                if (!Distribute.getInstance().notifyDownload(mReleaseDetails, intent)) {
                     AppCenterLog.info(LOG_TAG, "Release is downloaded. Starting to install it.");
                     Distribute.getInstance().setInstalling(mReleaseDetails);
                     Distribute.getInstance().showSystemSettingsDialogOrStartInstalling(downloadId, totalSize);